Check if backup exists before removing

+++ couchbase_backup.sh
@@ -56,7 +56,11 @@ configure () {
   aws configure set aws_secret_access_key ${AWS_SECRET_ACCESS_KEY}
   aws configure set region ${AWS_DEFAULT_REGION}
   
-  rm -r ${BACKUP_PATH}
+  if [ ! -d ${BACKUP_PATH} ]
+  then
+    rm -r ${BACKUP_PATH}
+  fi
+  
   mkdir -p ${BACKUP_PATH}
 }
